+++ packages/frontend/components/Forms/Form.module.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'styles/variables';
 
 $normal-font-size: 24px;
 $grey: #3f3f3f;
