+++ packages/frontend/components/Forms/FormVideo.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import TextInput from './TextInput'
 import CheckConfirm from './CheckConfirm'
 import PropTypes from 'prop-types'
-import styles from '../../styles/Form.module.scss'
+import styles from './Form.module.scss'
 
 interface dataType {
   creator: string
